% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_data.R
\docType{data}
\name{continuous_data}
\alias{continuous_data}
\title{Example continuous data}
\format{
A data frame with 1000 rows and 12 variables:
\describe{
\item{SBP Change}{The change in the systolic blood pressure (in mmHg) of patients over the 1-year study duration.}
\item{Statomycin (Y/N)}{\code{1} if the patient received statomycin, \code{0} otherwise.}
\item{Age Z-score}{The patient's standardized age.}
\item{Physical Activity Score}{The patient's standardized physical activity level, measured by questionnaire.}
\item{Diabetes (Y/N)}{\code{1} if the patient has diabetes, \code{0} otherwise.}
\item{Dietary Score}{The patient's standardized concordance with the Canada Food Guide, measured by questionnaire.}
\item{Air Quality Index}{The standardized air quality of the patient's hometown.}
\item{CAD (Y/N)}{\code{1} if the patient has coronary artery disease, \code{0} otherwise.}
\item{BMI Z-score}{The patient's BMI, relative to age and sex.}
\item{Stress Score}{The patient's day-to-day stress levels, measured by questionnaire.}
\item{Sex (M/F)}{\code{1} if the patient is male, \code{0} otherwise.}
\item{Happiness Score}{The patient's satisfaction with their life, measured by questionnaire.}
}
}
\usage{
data(continuous_data)
}
\description{
Data from a simulated study investigating the effects of a treatment, statomycin, on a continuous outcome, change in systolic blood pressure (SBP) over a 1-year period.
The data was simulated such that statomycin increases the expected SBP change over the period
by 4 mmHg for patients who were expected to change by less than 1 mmHg, and by 1 mm Hg for all other patients.
All continuous variables included were z-scored.
}
\keyword{datasets}
