% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_match_sets.R
\name{create_match_sets}
\alias{create_match_sets}
\title{Create Matched Sets Using Pair Matching}
\usage{
create_match_sets(treated, risk_scores, control_treated_ratio)
}
\arguments{
\item{treated}{A binary vector indicating treatment status (1 for treated, 0 for control).}

\item{risk_scores}{A numeric vector of risk scores for all individuals.}

\item{control_treated_ratio}{An integer specifying the ratio of controls to treated individuals for matching.}
}
\value{
A list containing the following elements:
\itemize{
\item{"matched_sets"}: The result of the pair matching process, containing matched sets of treated and control individuals.
\item{"patients_df"}: A data frame containing the treatment status, risk scores, and indices of all individuals.
}
}
\description{
This function creates matched sets of treated and control individuals using pair matching based on a risk score.
}
\details{
This function uses the \code{pairmatch()} function from the \code{optmatch} package to match treated and control individuals based on their risk scores. It supports specifying a control-to-treated ratio and removes any unmatched individuals from the results.
}
\author{
Danny Del Rosso, Erin Craig, Maksim Helmann
}
