% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sweetspot.R
\name{find_sweetspot}
\alias{find_sweetspot}
\title{Find Optimal Sweet Spot for Treatment Effects}
\usage{
find_sweetspot(
  treatment_effect,
  ntrials_significance = 1000,
  ntrials_bias = 1000,
  min_size_fraction = 1/20,
  max_size_fraction = 1/2
)
}
\arguments{
\item{treatment_effect}{Numeric vector containing the observed treatment effects.}

\item{ntrials_significance}{Integer. The number of resampling trials for calculating the p-value. Default is 1000.}

\item{ntrials_bias}{Integer. The number of resampling trials for debiasing the computed statistics. Default is 1000.}

\item{min_size_fraction}{Numeric. The minimum size of a window as a fraction of the total length of \code{treatment_effect}. Default is 1/20.}

\item{max_size_fraction}{Numeric. The maximum size of a window as a fraction of the total length of \code{treatment_effect}. Default is 1/2.}
}
\value{
A list with the following elements:
\describe{
\item{start_index}{The starting index of the optimal window with the highest statistic.}
\item{end_index}{The ending index of the optimal window with the highest statistic.}
\item{statistic}{The maximum deviation of the treatment effect in the optimal window from the global mean.}
\item{mean_inside}{The mean treatment effect for values inside the optimal window.}
\item{mean_outside}{The mean treatment effect for values outside the optimal window.}
\item{mean_inside_debiased}{The debiased mean treatment effect inside the optimal window, adjusted using bootstrap resampling.}
\item{mean_outside_debiased}{The debiased mean treatment effect outside the optimal window, adjusted using bootstrap resampling.}
\item{means}{A vector of bootstrapped means inside the optimal window, used for diagnostics.}
\item{means_outside}{A vector of bootstrapped means outside the optimal window, used for diagnostics.}
\item{start_indices}{The starting indices of windows sampled during debiasing.}
\item{end_indices}{The ending indices of windows sampled during debiasing.}
\item{bias_inside}{An estimate of the bias for the mean inside the optimal window.}
\item{bias_outside}{An estimate of the bias for the mean outside the optimal window.}
\item{p_value}{The p-value for the observed statistic, based on resampling.}
}
}
\description{
This function identifies the "sweet spot" in treatment effects by analyzing deviations from the global mean across potential window sizes. It computes statistics for all valid windows, evaluates statistical significance through resampling, and debiases the statistics using bootstrapping.
}
\author{
Danny Del Rosso, Erin Craig, Maksim Helmann
}
