% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sweetspot.R
\name{one_statistic}
\alias{one_statistic}
\title{Find Optimal Sweet Spot for Treatment Effects}
\usage{
one_statistic(k, benefits_cumsum, n, overall_mean, stat = "mean")
}
\arguments{
\item{treatment_effect}{A numeric vector containing the observed treatment effects.}

\item{ntrials_significance}{Integer. The number of resampling trials for calculating significance. Default is 100.}

\item{ntrials_bias}{Integer. The number of resampling trials for debiasing the computed statistics. Default is 100.}

\item{min_size_fraction}{Numeric. The minimum size of the window as a fraction of the total length. Default is 1/20.}

\item{max_size_fraction}{Numeric. The maximum size of the window as a fraction of the total length. Default is 1/2.}
}
\value{
A list with elements:
\describe{
\item{start_index}{The starting index of the optimal window.}
\item{end_index}{The ending index of the optimal window.}
\item{statistic}{The computed statistic for the optimal window.}
\item{mean_inside}{Mean of values inside the optimal window.}
\item{mean_outside}{Mean of values outside the optimal window.}
\item{mean_inside_debiased}{Debiased mean of values inside the window.}
\item{mean_outside_debiased}{Debiased mean of values outside the window.}
\item{means}{Debiased means across bootstrap samples inside the window.}
\item{means_outside}{Debiased means across bootstrap samples outside the window.}
\item{start_indices}{Start indices of windows sampled during debiasing.}
\item{end_indices}{End indices of windows sampled during debiasing.}
\item{bias_inside}{Bias estimate for the mean inside the window.}
\item{bias_outside}{Bias estimate for the mean outside the window.}
\item{p_value}{The p-value for the observed treatment effect window.}
}
}
\description{
This function identifies the optimal "sweet spot" in treatment effects by computing statistics across possible window sizes, assessing significance through resampling, and debiasing the identified window's statistics.
}
\author{
Danny Del Rosso, Maksim Helmann
}
