% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_95CI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Calculate Odds Ratio}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{Coefficinet of the predictor}

\item{se}{Standard error of the predictor}

\item{siglevel}{Significance level}

\item{roundto}{Number of decimal places to round to}
}
\description{
Calculate Odds Ratio and 95\% Confidence Interval
}
