% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quintiles.R
\name{plot_quintiles}
\alias{plot_quintiles}
\title{Plot Treatment Effect at Different Quintiles}
\usage{
plot_quintiles(result, title, pvalue = NULL)
}
\arguments{
\item{result}{The output of the "sweetspot" wrapper function.}

\item{title}{A character string specifying the title of the plot.}

\item{pvalue}{An inputted p-value for a user-performed significance test. Defaults to the p-value from sweetspot.}
}
\value{
A \code{ggplot} object representing the sweet spot visualization.
}
\description{
This function visualizes treatment effect estimates by splitting the patients into quintiles of increasing disease severity and computing the Average Treatment Effect (ATE) in each.
}
\examples{
\dontrun{
  # Assuming `result` contains model output and match data
  plot <- plot_sweetspot(result, "Quintile_Plot")

  print(plot)
}
}
\author{
Danny Del Rosso, Maksim Helmann
}
