% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_result.R
\name{plot_sweetspot}
\alias{plot_sweetspot}
\title{Visualize the sweet spot analysis on clinical trial data.}
\arguments{
\item{sweetspot_result}{The result from a sweetspot analysis.}

\item{title}{The title for your plot.}
}
\value{
A ggplot.
}
\description{
Visualize the sweet spot analysis on clinical trial data.
}
