% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_scores.R
\name{risk_scores}
\alias{risk_scores}
\title{Calculate risk scores from prognostic data}
\usage{
risk_scores(
  treated,
  covariates,
  response,
  family = "binomial",
  regularized = F,
  nfolds = NULL
)
}
\arguments{
\item{treated}{A binary vector indicating treatment status (1 for treated, 0 for control).}

\item{covariates}{A matrix with rows corresponding to individuals and colums to covariates to include in the mode;.}

\item{response}{A numeric vector of response values for the individuals.}

\item{family}{A string indicating the model family to be fit. Default is "binomial".}

\item{regularized}{later}

\item{nfolds}{later}
}
\description{
This function uses cross-validated elastic net regression to compute the predilection scores of patients from an inputted set of covariates
}
\author{
Danny Del Rosso, Maksim Helmann
}
