% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_norm_sweetspot.R
\name{sim_norm_sweetspot}
\alias{sim_norm_sweetspot}
\title{Simulate Continuous Sweetspot Data}
\usage{
sim_norm_sweetspot(
  n,
  magnitude,
  window = c(-1, 1),
  base_effect = 0.2,
  true_covs = 10,
  inc_covs = 10,
  random = F,
  multicoll = 0,
  xbin = T
)
}
\arguments{
\item{n}{A positive integer representing the amount of observations to be generated.}

\item{magnitude}{The additional increase in the expected value of the continuous outcome caused by the treatment in the sweetspot.}

\item{window}{A 2-element numeric vector with the boundaries of the outcome values that define the sweetspot. Default is \code{c(-1, 1)}.}

\item{base_effect}{The increase in the expected value of the continuous outcome caused by the treatment outside of the sweetspot. Default is \code{0.05}.}

\item{true_covs}{The number of covariates used to generate the true predilection probabilities. Default is \code{10}.}

\item{inc_covs}{The number of covariates included in the model. Default is \code{10} (i.e. all of the covariates that were used to generate the true probabilities).}

\item{random}{A boolean indicating if the predilection probabilities should be generated randomly, overriding all previous information about covariates. Default is \code{FALSE}.}

\item{multicoll}{A positive number indicating the degree of relatedness of the predictors. Ranges between \code{0} (no multicollinearity) to \code{1} (covariates in the model are largely controlled by other common, unseen covariates).}

\item{xbin}{A boolean indicating if every third covariate should be dichotomized to generate binary covariates.}
}
\value{
A list with the following 5 elements:
\describe{
\item{treated}{A binary vector representing control (\code{0}) or treated (\code{1}) status}
\item{outcome}{A binary vector representing the presence of an outcome (\code{1}), here representing a positive outcome.}
\item{p}{The probabilities used to generate the outcome.}
\item{covariates}{A matrix of included covariates, with the rows representing each observation.}
}
}
\description{
This function generates continuous data compatible with the package, for a user-specified sweetspot.
}
\author{
Danny Del Rosso, Maksim Helmann
}
