% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_quintiles.R
\name{split_quintiles}
\alias{split_quintiles}
\title{Separate Treatment Outcomes Into Quintiles}
\usage{
split_quintiles(result)
}
\arguments{
\item{result}{The output of the "sweetspot" wrapper function.}
}
\value{
A list of vectors at each quintile containing the treatment effects of the relevant patients.
}
\description{
This function allows the user to split the patients into quintiles based on baseline risk and examine the outcomes in each quintile.
}
\examples{
\dontrun{
  # Assuming `result` contains model output and match data
  quintiles <- split_quintiles(result)
  print(quintiles)
}
}
\author{
Danny Del Rosso, Maksim Helmann
}
