% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sweetspot_object.R
\name{summary.sweetspot_object}
\alias{summary.sweetspot_object}
\title{Summarize the results of a Sweetspot analysis}
\usage{
\S3method{summary}{sweetspot_object}(object)
}
\arguments{
\item{object}{An object of class \code{sweetspot_object}.}
}
\value{
The 6 key summary statistics of any sweetspot analysis.
}
\description{
Provide summary statistics for an object of class \code{sweetspot_object}.
}
