% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_quintiles.R
\name{test_quintiles}
\alias{test_quintiles}
\title{Test Significance of Effect Difference Across Quintiles}
\usage{
test_quintiles(result, nsim = 10000)
}
\arguments{
\item{result}{The output of the "sweetspot" wrapper function}

\item{nsim}{The number of simulations to use in calculating the p-value.}
}
\value{
The computed p-value
}
\description{
plot_quintiles This function exemplifies a Monte-Carlo simulation-based approach for determining if the difference in treatment effect between risk quintiles is statistically significant.
}
\examples{
\dontrun{
  # Assuming `result` contains model output and match data
  pval <- quintile_test(result, nsim = 100)
  print(pval)
}
}
\author{
Danny Del Rosso, Maksim Helmann
}
