% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sweetspot.R
\name{plot_sweetspot}
\alias{plot_sweetspot}
\title{Plot Sweet Spot for Treatment Effects}
\usage{
plot_sweetspot(model, title, hypothesis)
}
\arguments{
\item{model}{A model object extracted from the \code{result} object, containing
indices for the sweet spot and mean treatment effect estimates.}

\item{title}{A character string specifying the title of the plot.}

\item{hypothesis}{A character string representing the null hypothesis for
the statistical test.}
}
\value{
A \code{ggplot} object representing the sweet spot visualization.
}
\description{
Plot Sweet Spot for Treatment Effects
}
\details{
The plot's x-axis is labeled based on the family of the model (binomial, poisson, or other). Additionally, it annotates the start and end of the sweet spot region.
}
\section{Functions}{
\itemize{
\item \code{plot_sweetspot()}: function visualizes treatment effect estimates using a smoothed spline, highlighting regions inside and outside the sweet spot with both original and debiased treatment effects.

}}
\examples{
\dontrun{
  # Assuming `result` contains model output and match data
  plot <- plot_sweetspot(result$model, "Sweet Spot Analysis",
  "No sweet spot related to illness severity")
  print(plot)
}
}
\author{
Danny Del Rosso, Maksim Helmann
}
