% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example Data
Data on toys}
\format{
A dataframe
\describe{
 \item{toyid}{Toy ID}
 \item{toyname}{Toy Name}
 \item{price}{Price of the toy}
 \item{age}{Age group for the toy}
 }
}
\source{
Aya
}
\usage{
toydata
}
\description{
Example Data
Data on toys
}
\references{
My reference
}
\keyword{datasets}
